+++ plugins/azure/pkg/instance/costmanagement/costmanagement_test.go
@@ -1,0 +1,12 @@
+package costmanagement
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestNew(t *testing.T) {
+	client := New("", nil)
+	require.NotEmpty(t, client)
+}