+++ plugins/azure/pkg/instance/kubernetesservices/kubernetesservices.go
@@ -8,15 +8,21 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice"
 )
 
-// Client is the client to interact with the kubernetes services API.
-type Client struct {
+// Client is the interface for a client to interact with the Azure Kubernetes services api.
+type Client interface {
+	ListManagedClusters(ctx context.Context, resourceGroup string) ([]*armcontainerservice.ManagedCluster, error)
+	GetManagedCluster(ctx context.Context, resourceGroup, managedCluster string) (armcontainerservice.ManagedClustersGetResponse, error)
+	ListNodePools(ctx context.Context, resourceGroup, managedCluster string) ([]*armcontainerservice.AgentPool, error)
+}
+
+type client struct {
 	subscriptionID        string
 	managedClustersClient *armcontainerservice.ManagedClustersClient
 	agentPoolsClient      *armcontainerservice.AgentPoolsClient
 }
 
 // ListManagedClusters returns all managed clusters from the given resource group.
-func (c *Client) ListManagedClusters(ctx context.Context, resourceGroup string) ([]*armcontainerservice.ManagedCluster, error) {
+func (c *client) ListManagedClusters(ctx context.Context, resourceGroup string) ([]*armcontainerservice.ManagedCluster, error) {
 	var managedClusters []*armcontainerservice.ManagedCluster
 
 	pager := c.managedClustersClient.ListByResourceGroup(resourceGroup, &armcontainerservice.ManagedClustersListByResourceGroupOptions{})
@@ -32,12 +38,12 @@ func (c *Client) ListManagedClusters(ctx context.Context, resourceGroup string) 
 }
 
 // GetManagedCluster returns a single managed cluster.
-func (c *Client) GetManagedCluster(ctx context.Context, resourceGroup, managedCluster string) (armcontainerservice.ManagedClustersGetResponse, error) {
+func (c *client) GetManagedCluster(ctx context.Context, resourceGroup, managedCluster string) (armcontainerservice.ManagedClustersGetResponse, error) {
 	return c.managedClustersClient.Get(ctx, resourceGroup, managedCluster, &armcontainerservice.ManagedClustersGetOptions{})
 }
 
 // ListNodePools list all node pools for a manged cluster.
-func (c *Client) ListNodePools(ctx context.Context, resourceGroup, managedCluster string) ([]*armcontainerservice.AgentPool, error) {
+func (c *client) ListNodePools(ctx context.Context, resourceGroup, managedCluster string) ([]*armcontainerservice.AgentPool, error) {
 	var nodePools []*armcontainerservice.AgentPool
 
 	pager := c.agentPoolsClient.List(resourceGroup, managedCluster, &armcontainerservice.AgentPoolsListOptions{})
@@ -53,11 +59,11 @@ func (c *Client) ListNodePools(ctx context.Context, resourceGroup, managedCluste
 }
 
 // New returns a new client to interact with the kubernetes services API.
-func New(subscriptionID string, credentials *azidentity.ClientSecretCredential) *Client {
+func New(subscriptionID string, credentials *azidentity.ClientSecretCredential) Client {
 	managedClustersClient := armcontainerservice.NewManagedClustersClient(subscriptionID, credentials, &arm.ClientOptions{})
 	agentPoolsClient := armcontainerservice.NewAgentPoolsClient(subscriptionID, credentials, &arm.ClientOptions{})
 
-	return &Client{
+	return &client{
 		subscriptionID:        subscriptionID,
 		managedClustersClient: managedClustersClient,
 		agentPoolsClient:      agentPoolsClient,