+++ CHANGELOG.md
@@ -31,6 +31,7 @@ NOTE: As semantic versioning states all 0.y.z releases can contain breaking chan
 - [#270](https://github.com/kobsio/kobs/pull/270): [applications] Allow the usage of dashboards in application dependencies.
 - [#271](https://github.com/kobsio/kobs/pull/271): [applications] Add support for external applications.
 - [#272](https://github.com/kobsio/kobs/pull/272): [helm] Add Helm plugin, to manage Helm releases via kobs.
+- [#273](https://github.com/kobsio/kobs/pull/273): [azure] Add tests for Azure plugin.
 
 ### Fixed